+++ templates/error.php
@@ -13,7 +13,7 @@ script('fbsync', 'login');
 style('fbsync', 'fbsync');
 ?>
 
-<div id="app">
+<div id="app" class="fbsync">
 	<div id="app-navigation">
 		<?php print_unescaped($this->inc('part.navigation')); ?>
 		<?php print_unescaped($this->inc('part.settings')); ?>